IGNITE-19291 Fix review comment
