+++ modules/configuration/build.gradle
@@ -22,23 +22,6 @@ apply from: "$rootDir/buildscripts/java-test-fixtures.gradle"
 
 description = 'ignite-configuration'
 
-sourceSets {
-    defaultsGenerator {
-        compileClasspath += sourceSets.main.output
-        runtimeClasspath += sourceSets.main.output
-    }
-}
-
-configurations {
-    defaultsGeneratorImplementation.extendsFrom implementation
-    defaultsGenerator {
-        canBeConsumed = true
-        canBeResolved = false
-        // this configuration will share the same dependencies as the main one
-        extendsFrom implementation, runtimeOnly
-    }
-}
-
 dependencies {
     api project(':ignite-configuration-api')
 
@@ -61,14 +44,4 @@ dependencies {
     testFixturesImplementation libs.mockito.core
     testFixturesImplementation libs.typesafe.config
     testFixturesImplementation libs.hamcrest.core
-}
-
-tasks.register('defaultsJar', Jar) {
-    from sourceSets.defaultsGenerator.output
-    archiveAppendix = "generator"
-}
-
-artifacts {
-    // declare both jars to provide enough dependencies to the caller
-    defaultsGenerator defaultsJar, jar
 }
