+++ modules/runner/src/main/java/org/apache/ignite/internal/configuration/CompoundModule.java
