+++ modules/runner/src/main/java/org/apache/ignite/internal/configuration/NodeConfigReadException.java
@@ -18,18 +18,14 @@
 package org.apache.ignite.internal.configuration;
 
 import org.apache.ignite.lang.ErrorGroups.NodeConfiguration;
-import org.apache.ignite.lang.IgniteInternalException;
+import org.apache.ignite.lang.IgniteException;
 
 /**
- * Throws when node bootstrap configuration file failed to create with IO problem.
+ * Throws when node bootstrap configuration read failed with IO problem.
  */
-public class NodeConfigCreateException extends IgniteInternalException {
+public class NodeConfigReadException extends IgniteException {
 
-    public NodeConfigCreateException(String msg) {
-        super(NodeConfiguration.CONFIG_FILE_CREATE_ERR, msg);
-    }
-
-    public NodeConfigCreateException(String msg, Throwable cause) {
-        super(NodeConfiguration.CONFIG_FILE_CREATE_ERR, msg, cause);
+    public NodeConfigReadException(String msg, Throwable cause) {
+        super(NodeConfiguration.CONFIG_READ_ERR, msg, cause);
     }
 }
