+++ modules/runner/src/main/java/org/apache/ignite/internal/configuration/NodeConfigWriteException.java
@@ -18,14 +18,14 @@
 package org.apache.ignite.internal.configuration;
 
 import org.apache.ignite.lang.ErrorGroups.NodeConfiguration;
-import org.apache.ignite.lang.IgniteInternalException;
+import org.apache.ignite.lang.IgniteException;
 
 /**
- * Exception that gets thrown when a node bootstrap configuration file is malformed.
+ * Throws when writing to node bootstrap configuration failed with IO problem.
  */
-public class NodeConfigParseException extends IgniteInternalException {
+public class NodeConfigWriteException extends IgniteException {
 
-    public NodeConfigParseException(String msg, Throwable cause) {
-        super(NodeConfiguration.CONFIG_PARSE_ERR, msg, cause);
+    public NodeConfigWriteException(String message, Throwable cause) {
+        super(NodeConfiguration.CONFIG_WRITE_ERR, message, cause);
     }
 }
